predictionModel/requirements.txt
prophet
pandas
psycopg2-binary
python-dotenv